Check and update state.finished atomically
We may have data promises that lead to other data promises, which result
in multiple checks for `this.state.finished` when a page has finished
rendering. `this.setState` doesn't act immediately, so both checks will
see a false value, and we'll take page-finished actions twice. Instead,
we should check and set the state as an atomic operation.